--- a/dotcom-rendering/src/client/startup.ts
+++ b/dotcom-rendering/src/client/startup.ts
@@ -7,17 +7,13 @@ const measure = (name: string, task: () => Promise<void>): void => {
 	start();
 
 	task()
-		// You could use 'finally' here to prevent this duplication but finally isn't supported
-		// in Chrome 59 which is used by Cypress in headless mode so if you do it will
-		// break the Cypress tests on CI
-		// See: https://github.com/cypress-io/cypress/issues/2651#issuecomment-432698837
 		.then(() => {
-			end();
-			log('dotcom', `🥾 Booted ${name} in ${end()}ms`);
+			const duration = end();
+			log('dotcom', `🥾 Booted ${name} in ${duration}ms`);
 		})
 		.catch(() => {
-			end();
-			log('dotcom', `🤒 Failed to boot ${name} in ${end()}ms`);
+			const duration = end();
+			log('dotcom', `🤒 Failed to boot ${name} in ${duration}ms`);
 		});
 };