<!-- In this repo you can label a PR with the "PR Deployment" label to deploy the code to a publicly accessible url -->

## What does this change?

We want to test our ability to identify poor web page performance as it happens. We have added the following conditions as indicators, and will report to Ophan when two or more are met:

- First contentful paint (FCP) is over 2.4 seconds
  - For the first half of June 2023, this has 10.91% sensitivity and 99.92% specificity
  
- Time to first byte (TTFB) is over 2.4 seconds
  - For the first half of June 2023, this has 3.33% sensitivity and 100.00% specificity
  
- Ophan took more than 5s to boot or 20ms to execute
  - This is just a guess at approximating performance, as we do not currently record any performance marks.

- The total duration of long tasks is above 120ms
  - We hope this can be a good indicator of FID. Only supported in Chromium browsers

## Why?

We believe that the best approache to page performance is to adaptively respond to user conditions. We need to establish a baseline of how good our predictions of poor web page performance are.

We have opted for extremely high specificity as a starting point, but a different tradeoff may be needed in the future.

## Screenshots

![FCP as CWV indicator](https://github.com/guardian/dotcom-rendering/assets/76776/624e6640-f984-455c-8f0b-7f8a667668f3)

![TTFB as CWV indicator](https://github.com/guardian/dotcom-rendering/assets/76776/c56cbbbe-7795-4524-b1d6-bc1f850bf600)