@mxdvl 2023-06-13
We need to make sure we document this more widely as it’s changed the Ophan data (@guardian/open-journalism)